Split message from client
